chore: remove Evm trait
