+++ src/eth/executor/mod.rs
@@ -1,13 +1,13 @@
 mod evm;
+mod evm_input;
+mod evm_result;
 #[allow(clippy::module_inception)]
 mod executor;
 mod executor_config;
-mod revm;
 
 pub use evm::Evm;
-pub use evm::EvmExecutionResult;
-pub use evm::EvmInput;
+pub use evm_input::EvmInput;
+pub use evm_result::EvmExecutionResult;
 pub use executor::Executor;
 pub use executor::ExecutorStrategy;
 pub use executor_config::ExecutorConfig;
-pub use revm::Revm;
